+++ examples/src/app/ApplicationAnt.tsx
@@ -3,6 +3,7 @@ import {
     ActionFormButtonAnt,
     ActionSpanAnt,
     BoolAnt,
+    BoolSwitchAnt,
     ConfirmAnt,
     DatePickerAnt,
     LabelAnt,
@@ -108,6 +109,7 @@ export class ApplicationAnt extends React.Component<{ app?: Application }> {
                             <br />
                             <h3>BoolUI Component</h3>
                             <BoolAnt operation={application.testBool} />
+                            <BoolSwitchAnt operation={application.testBool} />
                             {application.showCheckbox && (
                                 <p>
                                     <br />