+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
+### Added
+- BoolSwitchAnt (and BoolSwitchFormAnt) for using toggle switch to set boolean value
+
 # v.Next (Current)
 
 # [v0.2.0-beta.55](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.55) (2021-01-21)